pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fc-nmap"
dynamic = ["version"]
description = 'Farcaster network mapping'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "vrypan", email = "vrypan@gmail.comm" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  #"farcaster-py>=0.1.3",
  "requests",
  "grpcio==1.59.0",
  "grpcio-tools==1.59.0",
  "protobuf==4.24.4",
  "ipaddress"
]

[project.urls]
Documentation = "https://github.com/vrypan/fc-nmap#readme"
Issues = "https://github.com/vrypan/fc-nmap/issues"
Source = "https://github.com/vrypan/fc-nmap"

[project.scripts]
fc-nmap = "fc_nmap.cli:fc_nmap"

[tool.hatch.version]
path = "src/fc_nmap/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/fc_nmap tests}"

[tool.coverage.run]
source_pkgs = ["fc_nmap", "tests"]
branch = true
parallel = true
omit = [
  "src/fc_nmap/__about__.py",
]

[tool.coverage.paths]
fc_nmap = ["src/fc_nmap", "*/fc-nmap/src/fc_nmap"]
tests = ["tests", "*/fc-nmap/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
